cuentaCorriente: cambiar float por double

+++ cuentaCorriente.cpp
@@ -1,6 +1,6 @@
 #include "cuentaCorriente.hpp"
 
-float CuentaCorriente::getLineaCredito() { return linea_credito; }
+double CuentaCorriente::getLineaCredito() { return linea_credito; }
 
 void CuentaCorriente::setLineaCredito(double lc) { linea_credito = lc; }
